scripts/requirements.txt
kafka-python>=2.0.0
aws-msk-iam-sasl-signer-python>=1.0.0